Package restructure
Restructured to be a package with subpackages system.
